+++ src/components/visual/button/button.types.ts
@@ -1,0 +1,10 @@
+export enum ButtonVariantEnum {
+  primary = 'primary',
+  secondary = 'secondary',
+  neutral = 'neutral',
+}
+
+export enum ButtonSizeEnum {
+  small = 'small',
+  large = 'large',
+}