+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[0.0.24](https://github.com/multiversx/mx-sdk-dapp-ui/pull/186)] - 2025-08-06
+
+- [Fixed the Storybook manually outputted Tailwind file to scan SCSS files](https://github.com/multiversx/mx-sdk-dapp-ui/pull/185)
+- [Removed debounce from sign transactions button](https://github.com/multiversx/mx-sdk-dapp-ui/pull/184)
+- [Added generic Button component](https://github.com/multiversx/mx-sdk-dapp-ui/pull/183)
+- [Added Storybook support](https://github.com/multiversx/mx-sdk-dapp-ui/pull/180)
+
 ## [[0.0.23](https://github.com/multiversx/mx-sdk-dapp-ui/pull/182)] - 2025-08-04
 
 - [Updated styles on transaction table](https://github.com/multiversx/mx-sdk-dapp-ui/pull/181)